Notes: - moved similar part of locators to Base level - added Search - completed test for create and validate Contact
