+++ src/test/java/elements/accountsPage/Input.java
@@ -1,16 +1,16 @@
-package elements.ContactsPage;
+package elements.accountsPage;
 
 import elements.BaseElement;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 public class Input extends BaseElement {
     WebDriver driver;
     WebDriverWait wait;
     String inputFieldLabel;
-    String inputFieldLocator = BaseContactsPageInputElementLocator + "input";
+    //    String inputFieldLocator = "//*[contains(@class, 'modal-body')]//span[text()='%s']/ancestor::div[contains(@class, 'uiInput')]//input";
+    String inputFieldLocator = BaseAccountPageInputElementLocator + "//input";
 
     public Input(WebDriver driver, String inputFieldLabel) {
         this.driver = driver;
@@ -22,8 +22,6 @@ public class Input extends BaseElement {
         System.out.println(String.format("Writing text '%s' into input with label %s", inputText, inputFieldLabel));
         //TO DO
         //JS SET STYLE TO HIGHLIGHT FIELD
-        WebElement element = driver.findElement(By.xpath(String.format(inputFieldLocator, inputFieldLabel)));
-        highlightElement(driver, element);
-        element.sendKeys(inputText);
+        driver.findElement(By.xpath(String.format(inputFieldLocator, inputFieldLabel))).sendKeys(inputText);
     }
 }
