+++ src/test/java/elements/contactsPage/Input.java
@@ -1,16 +1,16 @@
-package elements.AccountsPage;
+package elements.contactsPage;
 
 import elements.BaseElement;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 public class Input extends BaseElement {
     WebDriver driver;
     WebDriverWait wait;
     String inputFieldLabel;
-    //    String inputFieldLocator = "//*[contains(@class, 'modal-body')]//span[text()='%s']/ancestor::div[contains(@class, 'uiInput')]//input";
-    String inputFieldLocator = BaseAccountPageInputElementLocator + "//input";
+    String inputFieldLocator = BaseContactsPageInputElementLocator + "input";
 
     public Input(WebDriver driver, String inputFieldLabel) {
         this.driver = driver;
@@ -22,6 +22,8 @@ public class Input extends BaseElement {
         System.out.println(String.format("Writing text '%s' into input with label %s", inputText, inputFieldLabel));
         //TO DO
         //JS SET STYLE TO HIGHLIGHT FIELD
-        driver.findElement(By.xpath(String.format(inputFieldLocator, inputFieldLabel))).sendKeys(inputText);
+        WebElement element = driver.findElement(By.xpath(String.format(inputFieldLocator, inputFieldLabel)));
+        highlightElement(driver, element);
+        element.sendKeys(inputText);
     }
 }
